Don't alert on app-warm up issues
If metrics are over threshold, but with a very small sample size (1),
then don't include the small sample in the average.  If over a 5 minute
period, at least 3 of the minute read points don't contain values that
weren't excluded due to low volume, then don't alert.

This change is intended to remove false positives caused by the first
few requests a day that may occur during an application's warm-up